Add index.html instead of FileUploader main class
It's slightly awkward having 2 main classes in the complete
project because ony one of them can run from the command line
by default. Consequently a user cannot take the "complete" project
and follow the instructions in the README because the server
cannot be run from there.

This change adds an index.html to the application so it
can be driven from a browser instead.

N.B. this sample is broken with Boot RC4 (RC3 still fine)